+++ java/client/src/test/java/org/signal/libsignal/net/ChatServiceConnectTests.java
@@ -7,6 +7,7 @@ package org.signal.libsignal.net;
 
 import static org.junit.Assert.*;
 
+import java.util.Map;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -60,6 +61,24 @@ public class ChatServiceConnectTests {
     Assume.assumeNotNull(ENABLE_TEST);
 
     final Network net = new Network(Network.Environment.STAGING, USER_AGENT);
+    final Listener listener = new Listener();
+    var chat = net.connectUnauthChat(listener).get();
+    chat.start();
+    Void disconnectFinished = chat.disconnect().get();
+
+    ChatServiceException disconnectReason = listener.disconnectReason.get();
+    assertNull(disconnectReason);
+  }
+
+  @Test
+  public void testConnectUnauthH2() throws Exception {
+    // Use the presence of the environment setting to know whether we should
+    // make network requests in our tests.
+    final String ENABLE_TEST = TestEnvironment.get("LIBSIGNAL_TESTING_RUN_NONHERMETIC_TESTS");
+    Assume.assumeNotNull(ENABLE_TEST);
+
+    final Network net = new Network(Network.Environment.STAGING, USER_AGENT);
+    net.setRemoteConfig(Map.of("useH2ForUnauthChat", "true"), Network.BuildVariant.BETA);
     final Listener listener = new Listener();
     var chat = net.connectUnauthChat(listener).get();
     chat.start();